Lots of little things.
Did quite a bit of work on the code. Simplified and revised many
things, found a bug I can’t understand (noted in the code). Most
important is that I revised the mechanics for collision properties for
each object. Instead of “hazard” or  not, now have a property that
calls a function. We can choose what function is called, so hazards can
call a different function than neutral objects.